Fixed plurals of Russian and Ukrainian feminine nouns.
In Russian and Ukrainian, the word 'month', for example, has masculine
grammatical gender. The word 'minute' in both languages has feminine
grammatical gender, and thus it's inflected for case in a slightly different
way. The word 'hour' is feminine in Ukrainian and masculine in Russian.
A little error in one form of feminine plurals is fixed for Russian and
Ukrainian locales. Tests are edited to reflect this.