return NotImplemented instead of raising TypeError
TypeError will already be raised by Python, if both object
implementations  don't know how to handle the other type.

This allows other classes to implement a __rOperation__, even
when arrow doesn't know how to handle it with __Operation__ itself.

Python 2 doesn't raise TypeError if binary arithmetic operations return
NotImplemented, so we fixup this in __cmp__.

From the docs:

NotImplemented:
Special value which should be returned by the binary special methods
(e.g. __eq__(), __lt__(), __add__(), __rsub__(), etc.) to indicate that
the operation is not implemented with respect to the other type; may be
returned by the in-place binary special methods (e.g. __imul__(),
__iand__(), etc.) for the same purpose. Its truth value is true.

Binary arithmetic operations:
[...]
If one of those methods does not support the operation with the supplied
arguments, it should return NotImplemented.

Rich comparison methods:
A rich comparison method may return the singleton NotImplemented if it
does not implement the operation for a given pair of arguments.