Consolidate German locales and add missing 'week'/'weeks' localization (#768)
* Add week/weeks granularity to GermanLocale

Signed-off-by: Nils Philippsen <nils@tiptoe.de>

* Use English names for German locales consistently

Signed-off-by: Nils Philippsen <nils@tiptoe.de>

* Clarify why GermanBaseLocale.describe() is needed

In the course, dispatch `only_distance == False` to the base class and
only compute the special case when needed.

Signed-off-by: Nils Philippsen <nils@tiptoe.de>

* Consolidate SwissLocale into GermanBaseLocale

SwissLocale was effectively a copy of GermanLocale, with some attributes
missing. Both locales declared themselves responsible for 'de', making
it dependent on chance which one was chosen for localizing. Having only
one class responsible (i.e. the common base class) avoids having to
maintain both when extending, e.g. as it happened when 'week'/'weeks'
was added.

Signed-off-by: Nils Philippsen <nils@tiptoe.de>

* Fix get_locale_by_class_name() and test

Previously, this would raise a KeyError rather than a ValueError if a
non-existent class is requested. Additionally, add a test for this
function.

Signed-off-by: Nils Philippsen <nils@tiptoe.de>